Casting is not needed, but is good practice
Always cast negative values in unsigned types since in some cases bugs
might be hidden. See:

https://www.securecoding.cert.org/confluence/display/seccode/INT18-C.+Evaluate+integer+expressions+in+a+larger+size+before+comparing+or+assigning+to+that+size#INT18-C.Evaluateintegerexpressionsinalargersizebeforecomparingorassigningtothatsize-NoncompliantCodeExample%28size_t%29
(cherry picked from commit 6c03d55a60af061e3238608e0f902b79c9c9b6b0)

+++ cf-serverd/server_transform.c
@@ -764,7 +764,8 @@ static size_t DeRegexify(StrList **sl, const char *host)
         }
         if (strchr(name, '*') != NULL)
         {
-            return -1; /* Can't handle a * later than the preamble. */
+            /* Can't handle a * later than the preamble. */
+            return (size_t) -1;
         }
 
         if (name > host || NULL != strchr(host, '\\'))
@@ -781,7 +782,7 @@ static size_t DeRegexify(StrList **sl, const char *host)
                     if (p[0] != '.')
                     {
                         /* Regex includes a non-dot escape */
-                        return -1;
+                        return (size_t) -1;
                     }
                 }
 #if 0
